Fixing conversion warnings

+++ PoutineStyle/StatePauseMenu.cpp
@@ -8,9 +8,9 @@ StatePauseMenu::StatePauseMenu(GameManager* game)
 	w = gameManager->getWindowWidth();
 	h = gameManager->getWindowHeight();
 
-	playButton = Button(sf::Vector2i(w / 2 - 75, 1.5 * (h / 4) - 25), 150, 50, sf::Color::Red, L"Reprendre", gameManager->getRenderWindow());
-	backToMainMenu = Button(sf::Vector2i(w / 2 - 75, 2.5 * (h / 4) - 25), 150, 50, sf::Color::Red, L"Menu principal", gameManager->getRenderWindow());
-	quitButton = Button(sf::Vector2i(w / 2 - 75, 3.5 * (h / 4) - 25), 150, 50, sf::Color::Red, L"Quitter", gameManager->getRenderWindow());
+	playButton = Button(sf::Vector2i(int(w / 2 - 75), int(1.5 * (h / 4) - 25)), 150, 50, sf::Color::Red, L"Reprendre", gameManager->getRenderWindow());
+	backToMainMenu = Button(sf::Vector2i(int(w / 2 - 75), int(2.5 * (h / 4) - 25)), 150, 50, sf::Color::Red, L"Menu principal", gameManager->getRenderWindow());
+	quitButton = Button(sf::Vector2i(int(w / 2 - 75), int(3.5 * (h / 4) - 25)), 150, 50, sf::Color::Red, L"Quitter", gameManager->getRenderWindow());
 }
 
 StatePauseMenu::~StatePauseMenu() {}
@@ -52,7 +52,7 @@ void StatePauseMenu::draw(double deltaTime)
 
 	pause = sf::Text("Pause", font, 35);
 	pause.setFillColor(sf::Color::White);
-	pause.setPosition(sf::Vector2f(w / 2 - 57, 1 * (h / 4) - 25));
+	pause.setPosition(sf::Vector2f(w / 2 - 57.f, 1 * (h / 4) - 25.f));
 	gameManager->getRenderWindow()->draw(pause);
 
 	playButton.draw();
